Rename GoPRMaker module to go_pr_maker
